# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ALTIL Tenant, Governed AI and Communications API
  version: 1.1.0
  summary: Introsoft ALTIL multi-tenant AI SaaS platform
  description: >-
    Tenant onboarding, organization hierarchy, scoped API keys, license and quota
    enforcement, invoicing, governed AI inference and encrypted multi-channel communication.
    AI calls fail closed when tenant, application, key, billing or policy checks cannot pass.
servers:
  - url: https://api.introsoft.com
    description: Production API gateway
  - url: http://localhost:3005
    description: Local development
security:
  - ALTILKey: []
  - BearerSession: []
...
paths:
  /api/v1/applications:
    get:
      tags: [Developers]
      summary: List applications visible in the caller's organization scope
      description: Returns applications from organizations visible under the authenticated ALTIL session. Application creation currently creates an initial inference credential as part of the same workflow.
      security:
        - BearerSession: []
      responses:
        '200':
          description: Visible application records
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DeveloperApplication'}}
        '401': {description: ALTIL session required}
        '403': {description: Organization read permission is not available}
    post:
      tags: [Developers]
      summary: Create an application and its initial inference credential
      description: Requires SUPER_ADMIN or TENANT_ADMIN plus tenant.update and apikeys.create for the same target organization. The generated credential secret is returned once in the response. Only the enforced read:inference scope is grantable.
      security:
        - BearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 120}
                customerId: {type: string, description: Target organization; defaults to the authorized current organization}
                tenantId: {type: string, description: Legacy target-organization alias}
                appIdentifier: {type: string, pattern: '^[a-zA-Z0-9][a-zA-Z0-9._-]{0,63}$'}
                description: {type: string, maxLength: 2000}
                environment: {type: string, enum: [development, staging, production], default: production}
                status: {type: string, enum: [active, suspended, revoked], default: active}
                allowedCapabilities: {type: array, maxItems: 50, items: {type: string, maxLength: 100}}
                assignedPolicyIds: {type: array, maxItems: 50, items: {type: string, maxLength: 128}}
                parentApplicationId: {type: [string, 'null']}
                applicationType: {type: string, enum: [application, sub_application, function]}
                functionIdentifier: {type: string, maxLength: 100}
                rateLimitRpm: {type: integer, minimum: 1, maximum: 1000000, default: 120}
                quotaMonthlyRequests: {type: integer, minimum: 1, maximum: 1000000000, default: 50000}
                contactEmail: {type: string, format: email, maxLength: 254}
      responses:
        '201':
          description: Application and one-time credential response
          content:
            application/json:
              schema:
                type: object
                required: [application, apiKey]
                properties:
                  application: {$ref: '#/components/schemas/DeveloperApplication'}
                  apiKey: {$ref: '#/components/schemas/DeveloperCredentialIssuedOnce'}
        '400': {description: Invalid application or parent-application fields}
        '401': {description: ALTIL session required}
        '403': {description: Role, organization authority, or runtime-scope grant denied}
        '503': {description: Application or credential registry unavailable}
...
  /api/v1/api-keys:
    get:
      tags: [Developers]
      summary: List credential metadata visible in the caller's organization scope
      description: Secret and digest fields are omitted. Existing credential secrets cannot be revealed again.
      security:
        - BearerSession: []
      responses:
        '200':
          description: Visible credential metadata
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DeveloperCredentialMetadata'}}
        '401': {description: ALTIL session required}
        '403': {description: Organization read permission is not available}
    post:
      tags: [Developers]
      summary: Create an application-bound API credential
      description: Requires apikeys.create and SUPER_ADMIN or TENANT_ADMIN. The application must be active and belong to the authorized organization. Secret is returned once; only read:inference is currently grantable.
      security:
        - BearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId: {type: string, description: Target organization; defaults to the authorized current organization}
                appId: {type: string}
                name: {type: string, maxLength: 120}
                scopes: {type: array, minItems: 1, maxItems: 1, items: {type: string, enum: [read:inference]}}
                ipWhitelist: {type: array, maxItems: 100, items: {type: string, description: IPv4 address or CIDR}}
                expiresAt: {type: [string, 'null'], format: date-time}
                rateLimitRpm: {type: integer, minimum: 1, maximum: 1000000, default: 120}
                monthlyRequestLimit: {type: [integer, 'null'], minimum: 1, maximum: 1000000000}
                monthlySpendLimitUsd: {type: [number, 'null'], minimum: 0, maximum: 1000000000}
                billingMode: {type: string, enum: [included, prepaid, metered]}
      responses:
        '201':
          description: Credential and one-time secret response
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeveloperCredentialIssuedOnce'}
        '400': {description: Invalid credential restrictions or application binding}
        '401': {description: ALTIL session required}
        '403': {description: Role, organization authority, or requested scope denied}
        '503': {description: Credential registry unavailable; no credential was issued}
...
  /api/v1/logs:
    get:
      tags: [Developers]
      summary: Read audit and request activity within authorized organization scope
      description: >-
        Requires audit.read and an authorized audit role. Results for records without a known
        application-to-organization relationship are omitted for scoped callers. Stored prompt and
        response preview fields may be present; this handler does not establish policy-based
        redaction, so treat preview values as potentially sensitive.
      security:
        - BearerSession: []
      parameters:
        - {name: appId, in: query, required: false, schema: {type: string}}
        - {name: providerId, in: query, required: false, schema: {type: string}}
        - {name: status, in: query, required: false, schema: {type: string}}
        - {name: search, in: query, required: false, schema: {type: string}}
      responses:
        '200':
          description: Authorized audit records; preview fields may contain sensitive content
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DeveloperAuditEvent'}}
        '401': {description: ALTIL session required}
        '403': {description: Audit permission or role is not available}
...
  /api/v1/capabilities:
    get:
      tags: [Tenants and IAM]
      summary: Read the source-reviewed capability inventory
      description: >-
        Returns a deliberately partial source-reviewed list of routes, authorization requirements,
        scope, and implementation status. Availability is evaluated for the current ALTIL session
        where applicable. This checkout-based inventory does not prove production deployment;
        omitted entries are unassessed, not absent.
      security:
        - BearerSession: []
      responses:
        '200':
          description: Capability inventory for the authenticated identity
          content:
            application/json:
              schema:
                type: object
                required: [inventoryComplete, evidenceBasis, generatedAt, capabilities]
                properties:
                  inventoryComplete: {type: boolean, const: false}
                  evidenceBasis: {type: string}
                  generatedAt: {type: string, format: date-time}
                  capabilities:
                    type: array
                    items:
                      type: object
                      required: [id, domain, name, method, route, authentication, scope, status, availableToCurrentIdentity]
                      properties:
                        id: {type: string}
                        domain: {type: string}
                        name: {type: string}
                        description: {type: string}
                        method: {type: string}
                        route: {type: string}
                        authentication: {type: string, enum: [PUBLIC, ALTIL_SESSION, TENANT_API_KEY]}
                        requiredRoles: {type: array, items: {type: string}}
                        requiredPermissions: {type: array, items: {type: string}}
                        scope: {type: string, enum: [PUBLIC, PLATFORM, ORGANISATION, TENANT]}
                        status: {type: string, enum: [IMPLEMENTED, PARTIALLY_IMPLEMENTED, CONFIGURED, PLANNED, UNAVAILABLE, REQUIRES_AUTHORIZATION, NOT_ASSESSED]}
                        ui: {type: array, items: {type: string}}
                        source: {type: string}
                        limitations: {type: array, items: {type: string}}
                        availableToCurrentIdentity: {type: [boolean, 'null']}
        '401': {description: ALTIL session required}
...
  /api/v1/models:
    get:
      tags: [Governed AI]
      summary: List routeable OpenRouter models
      description: Returns only live ALTIL registry models routed through OpenRouter. API-key callers require the currently enforced `read:inference` scope. No upstream catalogue refresh is triggered by this read.
      security:
        - ALTILKey: []
        - BearerSession: []
      responses:
        '200':
          description: OpenAI-compatible model list from the ALTIL registry
          content:
            application/json:
              schema:
                type: object
                required: [object, data]
                properties:
                  object: {type: string, const: list}
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, object, created, owned_by]
                      properties:
                        id: {type: string}
                        object: {type: string, const: model}
                        created: {type: integer}
                        owned_by: {type: string}
                        name: {type: string}
                        context_length: {type: integer}
                        capabilities: {type: array, items: {type: string}}
        '401': {description: API key or session is required}
        '403': {description: Tenant scope or inference permission denied}
        '429': {description: Key rate limit reached}
        '503': {description: OpenRouter runtime is not configured}
...
  /v1/models:
    get:
      tags: [Governed AI]
      summary: Alias for GET /api/v1/models
      security:
        - ALTILKey: []
        - BearerSession: []
      responses:
        '200': {description: OpenAI-compatible model list from the ALTIL registry}
        '401': {description: API key or session required}
        '403': {description: Tenant or scope denied}
        '503': {description: OpenRouter runtime is not configured}
...
  /api/v1/communications/channels/{id}/test:
    post:
      tags: [Communications]
      summary: Test SMTP connection or Firebase service account credentials
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Connection verified}
        '422': {description: Provider test failed}
...
  /api/v1/billing/refunds:
    get:
      tags: [Billing]
      summary: List tenant refund requests and their lifecycle status
      responses:
        '200': {description: Tenant-scoped refund records}
    post:
      tags: [Billing]
      summary: Request a partial or full refund against a captured invoice payment
      description: A request reserves its amount against the captured balance and requires finance approval before any provider action.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentIntentId, amount, reason]
              properties:
                tenantId: {type: string}
                paymentIntentId: {type: string}
                amount: {type: number, exclusiveMinimum: 0}
                reason: {type: string, minLength: 10, maxLength: 1000}
      responses:
        '201': {description: Refund request recorded for finance approval}
        '400': {description: Invalid reason or amount exceeds remaining captured balance}
        '403': {description: Payment belongs to another tenant}
...
  /api/v1/billing/schedules:
    get:
      tags: [Billing]
      summary: List recurring or threshold-based collection schedules
      responses:
        '200': {description: Tenant-scoped schedule metadata}
    post:
      tags: [Billing]
      summary: Create a separately consented monthly or usage-threshold debit schedule
      description: Usage threshold automation currently requires USD metering and a Stripe mandate. PayFast token charges are ZAR-only. iKhokha supports hosted one-time checkout in the current integration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentMethodId, scheduleType, consent, consentText]
              properties:
                tenantId: {type: string}
                paymentMethodId: {type: string}
                scheduleType: {type: string, enum: [threshold, monthly]}
                thresholdAmount: {type: number, exclusiveMinimum: 0}
                currency: {type: string, pattern: '^[A-Z]{3}$', description: Configured transaction currency; posted source currency is preserved}
                consent: {type: boolean, const: true}
                consentText: {type: string, minLength: 100}
      responses:
        '201': {description: Schedule created with a consent record}
        '400': {description: Mandate, provider, currency or threshold constraints failed}
...
  /api/v1/billing/accounting/summary:
    get:
      tags: [Billing]
      summary: Tenant-scoped receivables, chart-of-account and monthly journal rollups
      responses:
        '200': {description: Currency-preserving finance dashboard data}
        '503': {description: Accounting migration unavailable}
  /api/v1/billing/accounting/journals:
    get:
      tags: [Billing]
      summary: List tenant journals and debit/credit lines
      responses:
        '200': {description: Posted, immutable journal lines}
    post:
      tags: [Billing]
      summary: Post a balanced admin journal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, currency, description, lines]
              properties:
                tenantId: {type: string}
                currency: {type: string, pattern: '^[A-Z]{3}$', description: Configured transaction currency; posted source currency is preserved}
                description: {type: string}
                lines: {type: array, minItems: 2, maxItems: 100, items: {type: object, required: [accountCode, accountName, debit, credit]}}
      responses:
        '201': {description: Balanced journal posted}
        '422': {description: Journal is unbalanced or storage is unavailable}
...
components:
  securitySchemes:
    ALTILKey: {type: apiKey, in: header, name: x-api-key}
    BearerSession: {type: http, scheme: bearer, bearerFormat: ALTIL session}
    DeviceSecret: {type: apiKey, in: header, name: x-device-secret}
  schemas:
    DeveloperApplication:
      type: object
      required: [id, name, status, environment, allowedCapabilities, createdAt, updatedAt]
      properties:
        id: {type: string}
        customerId: {type: [string, 'null']}
        customerName: {type: string}
        parentApplicationId: {type: [string, 'null']}
        applicationType: {type: string, enum: [application, sub_application, function]}
        functionIdentifier: {type: [string, 'null']}
        appIdentifier: {type: string}
        name: {type: string}
        description: {type: string}
        status: {type: string, enum: [active, suspended, revoked]}
        environment: {type: string, enum: [development, staging, production]}
        allowedCapabilities: {type: array, items: {type: string}}
        rateLimitRpm: {type: integer}
        quotaMonthlyRequests: {type: integer}
        quotaUsedRequests: {type: integer}
        assignedPolicyIds: {type: array, items: {type: string}}
        contactEmail: {type: string}
        createdAt: {type: string}
        updatedAt: {type: string}
...